Move of delegations verify_role from schema/mod.rs to schema/verify.rs
